requirements.txt
coverage==4.5.3
Flask==2.0.2
mock==2.0.0
nose==1.3.7
numpy==1.21.5
pandas==1.3.5
PyYAML==6.0
pyyml==0.0.2
requests==2.21.0


